:art: Merge viewCtrl and view
Mostly shared responsibility, only using a separate flowtype definition
for the side-effects usage to not know and use streams directly from
the view.
