+++ decls/textual-velocity/_imports.js
@@ -1,0 +1,3 @@
+// imported packages used by most declared types
+import fs from 'fs'
+import Bacon from 'baconjs'
