+++ lib/session.js
@@ -3,7 +3,6 @@
 import PathWatcherFactory from './path-watcher-factory'
 import Interactor from './interactor'
 import Presenter from './presenter'
-import ReactView from './react-view'
 import SideEffects from './side-effects'
 import ViewCtrl from './view-ctrl'
 
@@ -18,21 +17,19 @@ export default class Session {
     this._panel = atom.workspace.addTopPanel({
       item: document.createElement('div')
     })
-    this._view = new ReactView(this._panel)
-    this._viewCtrl = new ViewCtrl(this._view)
+    this._viewCtrl = new ViewCtrl(this._panel)
 
     const pathWatcherFactory = new PathWatcherFactory(service)
     const interactor = new Interactor(this._viewCtrl, pathWatcherFactory, service)
     const presenter = new Presenter(interactor, service.columnsP)
-    this._atomSideEffects = new SideEffects(this._panel, this._view, presenter, service)
+    this._atomSideEffects = new SideEffects(this._panel, this._viewCtrl, presenter, service)
 
     this._viewCtrl.activate()
   }
 
   dispose () {
-    this._viewCtrl.deactivate()
+    this._viewCtrl.dispose()
     this._atomSideEffects.dispose()
-    this._view.dispose()
     this._panel.destroy()
   }
 }
