+++ lib/side-effects.js
@@ -14,20 +14,21 @@ export default class SideEffects {
     this._panel = panel
     const togglePanelCmdS = atoms.createCommandStream('atom-workspace', 'textual-velocity:toggle-panel')
     const toggleAtomWinCmdS = atoms.createCommandStream('atom-workspace', 'textual-velocity:toggle-atom-window')
+    const sortS = presenter.sortP.changes()
 
     this._disposables = new DisposableValues(
-      this._saveListHeightOnPanelResize(view.listHeightS),
-      this._saveSortDirectionOnChange(view),
-      this._saveSortFieldOnChange(view),
-      this._saveRowHeightOnWinResize(presenter),
+      this._saveListHeightOnPanelResize(presenter.listHeightP.changes()),
+      this._saveSortDirectionOnChange(sortS.map('.direction')),
+      this._saveSortFieldOnChange(sortS.map('.field')),
+      this._saveRowHeightOnWinResize(presenter.rowsS),
       this._focusTextInputOnChanges(toggleAtomWinCmdS, togglePanelCmdS),
-      this._toggleAtomWinOnCmd(panel, toggleAtomWinCmdS),
-      this._togglePanelOnCmd(panel, togglePanelCmdS),
+      this._toggleAtomWinOnCmd(toggleAtomWinCmdS, panel),
+      this._togglePanelOnCmd(togglePanelCmdS, panel),
       this._renderLoadingOnLoadingEvent(presenter, view),
       this._renderResultsOnDataChanges(presenter, view),
-      this._previewSelectedPath(presenter),
-      this._openTextEditorOnOpenEvent(presenter),
-      this._saveEditedCellOnSave(presenter, service.fileWritersP),
+      this._previewSelectedPath(presenter.selectedPathS),
+      this._openTextEditorOnOpenEvent(presenter.openPathS),
+      this._saveEditedCellOnSave(presenter.saveEditedCellContentS, service.fileWritersP),
       this._updateConfigSchemaOnColumnsChange(service.columnsP)
     )
   }
@@ -48,26 +49,24 @@ export default class SideEffects {
       })
   }
 
-  _saveSortDirectionOnChange (view: ViewType) {
-    return view
-      .sortDirectionS
+  _saveSortDirectionOnChange (sortDirectionS: Bacon.Stream) {
+    return sortDirectionS
       .onValue((direction: string) => {
         atom.config.set('textual-velocity.sortDirection', direction)
       })
   }
 
-  _saveSortFieldOnChange (view: ViewType) {
-    return view
-      .sortFieldS
+  _saveSortFieldOnChange (sortFieldS: Bacon.Stream) {
+    return sortFieldS
       .onValue((field: string) => {
         atom.config.set('textual-velocity.sortField', field)
       })
   }
 
-  _saveRowHeightOnWinResize (presenter: PresenterType) {
+  _saveRowHeightOnWinResize (rowsS: Bacon.Stream) {
     return Bacon
       .mergeAll(
-        presenter.rowsS.take(1),
+        rowsS.take(1),
         Bacon.fromEvent(window, 'resize'))
       .debounceImmediate(50) // ms
       .onValue(() => {
@@ -95,7 +94,7 @@ export default class SideEffects {
       })
   }
 
-  _togglePanelOnCmd (panel: Atom$Panel, togglePanelCmdS: Bacon.Stream) {
+  _togglePanelOnCmd (togglePanelCmdS: Bacon.Stream, panel: Atom$Panel) {
     return togglePanelCmdS
       .onValue(() => {
         if (panel.isVisible()) {
@@ -106,7 +105,7 @@ export default class SideEffects {
       })
   }
 
-  _toggleAtomWinOnCmd (panel: Atom$Panel, toggleAtomWinCmdS: Bacon.Stream) {
+  _toggleAtomWinOnCmd (toggleAtomWinCmdS: Bacon.Stream, panel: Atom$Panel) {
     return toggleAtomWinCmdS
       .onValue(() => {
         if (atom.getCurrentWindow().isFocused()) {
@@ -154,8 +153,8 @@ export default class SideEffects {
       })
   }
 
-  _previewSelectedPath (presenter: PresenterType) {
-    return presenter.selectedPathS
+  _previewSelectedPath (selectedPathS: Bacon.Stream) {
+    return selectedPathS
       .debounce(50) // ms
       .diff(undefined, (prevFile, selectedFile) => [prevFile, selectedFile])
       .onValue(([oldPath, newPath]) => {
@@ -174,17 +173,17 @@ export default class SideEffects {
       })
   }
 
-  _openTextEditorOnOpenEvent (presenter: PresenterType) {
-    return presenter.openPathS
+  _openTextEditorOnOpenEvent (openPathS: Bacon.Stream) {
+    return openPathS
       .onValue((path: string) => {
         atom.workspace.open(path)
       })
   }
 
-  _saveEditedCellOnSave (presenter: PresenterType, fileWritersP: Bacon.Property) {
+  _saveEditedCellOnSave (saveEditedCellContentS: Bacon.Stream, fileWritersP: Bacon.Property) {
     return fileWritersP
-      .combine(presenter.saveEditedCellContentS, R.pair)
-      .sampledBy(presenter.saveEditedCellContentS)
+      .combine(saveEditedCellContentS, R.pair)
+      .sampledBy(saveEditedCellContentS)
       .onValue(([fileWriters, {editCellName, path, str}]) => {
         fileWriters
           .find(fw => fw.editCellName === editCellName)
